Turned `poketerm` and `resources` into modules with empty `__init__.py` files. Moved files around. Adjusted imports as needed.
